adding changes including driver start
